LibTTF: Split up Font.h into smaller internal header files.
